util_cpack: Add support for non 16-bit sample widths
The util_cpack module currently assumes a sample width of 16-bit. Using it
with signals that use a different sample width results in incorrect
behavior.

Replace all the hard-coded constant 16 and related values with a
configurable module parameter. This allows the channel pack module to be
used for sample widths other than 16.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>